Modularize main teams page routes in separate file
